+++ Package/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.4
 - Added Consent Handlers to internal components to fix a crashing bug
 - Adding number pad component
+- Created NotificationOverlayView to manage notificationMaster.
 
 ## 0.0.3 
 - Audited FloatingContainerElements to be more consistent